+++ apps/backend/app/controllers/api/v1/pipelines_controller.rb
@@ -4,17 +4,23 @@ module Api
   module V1
     class PipelinesController < Api::BaseController
       def index
-        pipelines = current_account.pipelines.order(:created_at)
-        render json: { pipelines: pipelines.map { |p| PipelineSerializer.new(p).serializable_hash } }
+        scope = current_account.pipelines.order(:created_at)
+        @pagy, records = pagy(scope, items: per_page)
+        render json: {
+          pipelines: records.map { |p| PipelineSerializer.new(p).serializable_hash },
+          pagination: pagy_meta(@pagy)
+        }
       end
 
       def show
         pipeline = current_account.pipelines.find(params[:id])
+        authorize pipeline
         render json: { pipeline: PipelineSerializer.new(pipeline).serializable_hash }
       end
 
       def create
         pipeline = current_account.pipelines.new(pipeline_params)
+        authorize pipeline
         if pipeline.save
           render json: { pipeline: PipelineSerializer.new(pipeline).serializable_hash }, status: :created
         else
@@ -24,6 +30,7 @@ module Api
 
       def update
         pipeline = current_account.pipelines.find(params[:id])
+        authorize pipeline
         if pipeline.update(pipeline_params)
           render json: { pipeline: PipelineSerializer.new(pipeline).serializable_hash }
         else
@@ -33,6 +40,7 @@ module Api
 
       def destroy
         pipeline = current_account.pipelines.find(params[:id])
+        authorize pipeline
         pipeline.destroy
         head :no_content
       end
@@ -41,6 +49,14 @@ module Api
 
       def pipeline_params
         params.require(:pipeline).permit(:name, :description, :status, :primary, stage_definitions: [:name])
+      end
+
+      def per_page
+        (params[:per_page].presence || Pagy::DEFAULT[:items]).to_i
+      end
+
+      def pagy_meta(p)
+        { page: p.page, items: p.items, count: p.count, pages: p.pages }
       end
     end
   end
